+++ assets/js/utils/form/select-advanced.js
@@ -1,0 +1,5 @@
+import SelectManager from "./SelectManager";
+
+document.querySelectorAll('select[data-select="advanced"]').forEach(selectElt => {
+    new SelectManager('#' + selectElt.id)
+})